mvc architecture applied
