+++ lib/main.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:money_tracker/models/category/category_model.dart';
 import 'package:money_tracker/models/transaction/transaction_model.dart';
-import 'package:money_tracker/provider/category_provider.dart';
-import 'package:money_tracker/provider/transaction_provider.dart';
-import 'package:money_tracker/screens/splash_screen.dart';
+import 'package:money_tracker/controller/category_controller.dart';
+import 'package:money_tracker/controller/transaction_controller.dart';
+import 'package:money_tracker/view/splash_screen.dart';
 import 'package:provider/provider.dart';
 
 const setkey = "";
