+++ lib/widgets/bottom_navbar.dart
@@ -1,9 +1,9 @@
 import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:flutter/material.dart';
 
-import 'package:money_tracker/screens/category/category_screen.dart';
-import 'package:money_tracker/screens/home/home_screen.dart';
-import 'package:money_tracker/screens/profile/profile_screen.dart';
+import 'package:money_tracker/view/category/category_screen.dart';
+import 'package:money_tracker/view/home/home_screen.dart';
+import 'package:money_tracker/view/profile/profile_screen.dart';
 
 class BottomNavBar extends StatefulWidget {
   const BottomNavBar({super.key, required this.text});
